Rename testers2 -> tester
