+++ src/main/scala/chisel3/tester/FirrterpreterBackend.scala
@@ -1,6 +1,6 @@
 // See LICENSE for license details.
 
-package chisel3.testers2
+package chisel3.tester
 
 import chisel3._
