+++ src/test/scala/chisel3/tests/BasicTest.scala
@@ -3,7 +3,7 @@ package chisel3.tests
 import org.scalatest._
 
 import chisel3._
-import chisel3.testers2._
+import chisel3.tester._
 
 class BasicTest extends FlatSpec with ChiselScalatestTester {
   behavior of "Testers2"
