+++ src/test/scala/chisel3/tests/QueueTest.scala
@@ -4,8 +4,8 @@ import org.scalatest._
 
 import chisel3._
 import chisel3.util._
-import chisel3.testers2._
-import chisel3.testers2.TestAdapters._
+import chisel3.tester._
+import chisel3.tester.TestAdapters._
 
 class QueueTest extends FlatSpec with ChiselScalatestTester {
   behavior of "Testers2 with Queue"
